+++ tests/cases/source/error/try_finally.py
@@ -1,0 +1,15 @@
+try:
+    import os
+finally:
+    pass
+
+try:
+    try:
+        pass
+    finally:
+        pass
+    import ujson as json
+except ImportError:
+    import json
+
+print(json)